Refactor: 컴포넌트에서 페이지 디렉토리로 변경
